Add a minimal test suite for the Server class.
This only tests the connect methods as these are the only methods that do things
that might be incorrect independently of the server. Testing other methods would
be effectively be merely testing the remote server.

The connect methods need to be tested to ensure that two connections to the same
remote server return the same Server object.