New OutputPort API.
This API tries to better encapsulate the horrors of workflow output ports. This
version no longer relies on the TreeList implementation but uses a unified
PortValue type that can represent data, a list or an error with the PortData,
PortList or PortError classes, respectively.

The uk.org.taverna.server.client.utils package is now redundant and has been
deleted. Its sole classes were TreeList, an associated exception and unit tests.