Refactor connections code to handle streaming downloads.
An AbstractConnection class is also introduced to help keep connection
implementations cleaner in terms of code required. This simply abstracts away
some of the more generalised versions of the Connection interface so that a
minimal set of methods must be implemented for full functionality.

In addition all methods defined in the Connection interface now return something
to indicate success. If the method was previously void then it now returns a
boolean true or false.