Mark a Run as deleted when its delete() method is called.
If the Run.delete() method is called then the remote run instance is destroyed
but, of course, the local Run object persists. This change simply marks it as
deleted so that no more communication will be attempted to the server regarding
it.

Any cached data within the object should be preserved for further reference but
there is no guarantee as to what will have been cached.