Connection.update() methods now return URIs.
These methods used to return boolean to indicate success, which was fine when
they were only used to set simple parameters, but in future they will also be
used to upload larger amounts of data. This use use of the update methods will
sometimes be creating resources on the remote server, so the URI of these
created resources should be returned. For the old usage the URI of the parameter
that has been changed is returned.