Major refactoring of the server response exceptions.
A new common superclass for response exceptions is used to unify the behaviour.
This exception stores the URI for which resource access was problematic and
makes it available for retrieval. It also stores the server response code.

All the server response exceptions are also moved into the connection package as
that is the only place they are thrown from (by design).

Documentation is also brought up to date.